requirements.txt
feast==0.36.0
pyspark==3.4.1
delta-spark==2.4.0
feast-spark==0.36.0
python-dotenv==1.0.0